Import of GSLONDON/Tk-ListboxDnD-1.3 from CPAN.
gitpan-cpan-distribution: Tk-ListboxDnD
gitpan-cpan-version:      1.3
gitpan-cpan-path:         GSLONDON/Tk-ListboxDnD-1.3.tar.gz
gitpan-cpan-author:       GSLONDON
gitpan-cpan-maturity:     released

+++ Tk/ListboxDnD.pm
@@ -23,7 +23,7 @@ use strict;
 use warnings;
 
 use vars qw($VERSION);
-$VERSION = '1.100';
+$VERSION = '1.300';
 
 use base  qw(Tk::Derived Tk::Listbox);
 use Tk::widgets qw(Listbox );
@@ -36,24 +36,17 @@ sub Populate
 
     my($listbox, $args) = @_;
 
-    my $format = delete($args->{'-dragFormat'});
-    if(defined($format))
-      {
-	$listbox->{'ListboxDnD::dragFormat'} = $format;
-      }
-    else
-      {
-	$listbox->{'ListboxDnD::dragFormat'} = "<- %s ";
-      }
-
     $listbox->SUPER::Populate($args);
 
-    $listbox->Advertise('listbox' => $listbox );
-    $listbox->ConfigSpecs(DEFAULT => [$listbox]);
-    $listbox->Delegates(DEFAULT => $listbox);
+    $listbox->ConfigSpecs(
+        -dragformat => [qw/PASSIVE dragFormat DragFormat/, '<- %s'],
+    );
+
+    my (@bindtags) = $listbox->bindtags;
+    $listbox->bindtags([@bindtags[1, 0, 2, 3]]);
 
     ########################################################
-    # use button  3 to drag and drop the order of selected entries.
+    # use button  1 to drag and drop the order of selected entries.
     ########################################################
 
     my $dragging_text;
@@ -61,18 +54,19 @@ sub Populate
     my $moving_callback = sub {return};
     my $marker_index;
     my $marker_text;
+    my $have_moved;
 
-    # pressing button 3 selects the nearest element
+    # pressing button 1 selects the nearest element
     $listbox->bind
-      (       '<ButtonPress-3>' => 
+      (       '<ButtonPress-1>' => 
 	      sub
 	      { 
 		$marker_index = $listbox->nearest($Tk::event->y);
 		$dragging_text = $listbox->get($marker_index);
-		$marker_text = sprintf( $listbox->{'ListboxDnD::dragFormat'}, $dragging_text );
-		$listbox->delete($marker_index);
-		$listbox->insert($marker_index, $marker_text);
+		$marker_text = 
+			sprintf( $listbox->cget(-dragformat), $dragging_text );
 		
+		$have_moved = 0;
 		$moving_callback = sub 
 		  {
 		    my $current_index = $listbox->nearest($Tk::event->y);
@@ -86,15 +80,25 @@ sub Populate
 	      }
       );
 
-    # moving mouse while pressing button 3 shows where item will go
-    $listbox->bind( '<Motion>' => sub { &$moving_callback; } );
-
-    # releasing button 3 inserts the moving selection to the current index
+    # moving mouse while pressing button 1 shows where item will go
+    # note: in extended mode, with multiple items <CTL>-selected,
+    # dragging and dropping an item across selected items will
+    # unselect them, UNLESS the Tk::break is called at end of callback.
     $listbox->bind
-      (       '<ButtonRelease-3>' => 
+	( '<Motion>' => 
+	sub { $have_moved = 1; &$moving_callback; Tk::break; } 
+	);
+
+    # releasing button 1 inserts the moving selection to the current index
+    $listbox->bind
+      (       '<ButtonRelease-1>' => 
 	      sub
 	      { 
 		$moving_callback = sub {return;};
+		return unless $have_moved;
+		$listbox->delete($marker_index);
+		$listbox->insert($marker_index, $marker_text);
+
 		return unless(defined($dragging_text));
 		$listbox->delete($marker_index);
 		$listbox->insert($marker_index, $dragging_text);
@@ -116,21 +120,11 @@ __END__
 =head1 DESCRIPTION
 
    The intent is to add Drag and Drop functionality to the Tk::Listbox
-   widget.I would like some beta-testers to see if they can break
+   widget. You can drag items within the listbox to another location 
+   within the listbox by using <Button-1>.
+
+   I would like some beta-testers to see if they can break
    this module or find any bugs with it.
-
-   You can drag items within the listbox to another location 
-   within the listbox.
-
-   Issues: I would rather use Button-1, but that seems to conflict
-   with the binding that selects items in the listbox.
-
-   Button-3 does not seem to have this issue, although it seems
-   a little less intuitive to bind to that button.
-
-   oh well.
-
-
 
 =head2 EXPORT
 
@@ -144,7 +138,7 @@ __END__
         use Tk;
         use Tk::ListboxDnD;
         my $top = MainWindow->new();
-        my $listbox = $top->ListboxDnD(-dragFormat=>"<%s>")->pack();
+        my $listbox = $top->ListboxDnD(-dragformat=>"<%s>")->pack();
         $listbox->insert('end', qw/alpha bravo charlie delta echo fox/);
         MainLoop();
